requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cohere==5.18.0
comm==0.2.3
cssselect2==0.8.0
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
fastavro==1.12.0
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.3
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-cohere==0.4.6
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.28
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==1.26.4
openai==1.107.3
orjson==3.11.1
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
parso==0.8.5
pexpect==4.9.0
pgvector==0.3.6
pillow==10.4.0
platformdirs==4.4.0
playwright==1.55.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.10
psycopg-pool==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
pypdf==6.0.0
pyphen==0.17.2
pyright==1.1.405
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.0
ruff==0.13.2
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
tenacity==8.5.0
tiktoken==0.10.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zopfli==0.2.3.post1
zstandard==0.25.0
